posts: don't allow parent-child cycles.
Fix a bug that allowed posts to have cycles in parent-child relations. It wasn't possible for a post
to be its own parent or grandparent, but it was possible to make a post its own great-grandparent.
You could do this by making post p1 the parent of post p2, and p2 the parent of p3, and p3 the parent of p1.