docker-compose: upgrade to Postgres 16.1 (read before upgrading)
Upgrade Postgres from 12.3 to 16.1 in the Docker Compose file.

Fixes a bug introduced in 5eb0d1e18 where Danbooru wouldn't boot because the `gen_random_uuid()`
Postgres function was missing. GoodJob 3.22 uses this function, which is included by default in
Postgres 13 and up, but which requires the pgcrypto extension in Postgres 12.

Important: if you're using the Docker Compose file, you must follow the steps below to upgrade your
installation to Postgres 16. This will involve backing up and restoring your database.

Run the following commands to upgrade your database:

    # Stop the application.
    docker compose stop

    # Backup the database.
    docker compose run --rm -T danbooru pg_dumpall -d postgres://danbooru@danbooru-postgres-1 --clean --if-exists --verbose > danbooru-backup.pg_dump

    # Stop the database container (this won't delete any data)
    docker compose stop
    docker compose rm

    # Delete your old database. Important: make sure the backup was successful first!
    # The old database has to be deleted before upgrading because Postgres 16.1 can't read it.
    docker volume rm danbooru-data

    # Upgrade to the latest version of Danbooru. This will upgrade from Postgres 12.3 to Postgres 16.1.
    docker compose pull

    # Start the new version of Postgres.
    docker compose up -d postgres

    # Restore your database from backup.
    docker compose run --rm -T danbooru psql postgres://danbooru@danbooru-postgres-1 < danbooru-backup.pg_dump

    # Start the application again.
    docker compose up

You must do the above *before* you upgrade to this commit with `docker compose pull`. Otherwise the
new version of Postgres won't be able to read your database, and you'll have to downgrade back to
the previous version of Danbooru so you can follow the steps above to upgrade your database.