uploads: fail upload if autotagger API call fails.
Fail the upload if the autotagger service is enabled but the API call fails. In production there
were a few assets with missing AI tags because the autotagger service failed for whatever reason,
probably because it was briefly down or overloaded. In that case it's better to fail the upload and
let the user try again than to have random assets silently missing AI tags.

People running their own instances shouldn't enable the autotagger if they're not running it.