uploads: potential fix for uploads getting stuck.
Potential fix for uploads sometimes getting stuck in production and background jobs needing to be
restarted to unstick them.

When a media asset is processed, each thumbnail size is generated in parallel. Thumbnails are
additionally saved to each image server in parallel. This is done by calling `parallel_each`, which
creates `Concurrent::Promise`s. Promises are worked by a background thread pool. When a promise was
created, if there weren't any idle threads to work it immediately, then it would be queued up and
worked as soon as a thread became available. This meant that if we had too many uploads at once,
each of them would spawn a bunch of promises to generate thumbnails, and it was possible for these
promises to be queued up and interleaved in such a way that promise A was waiting for promise B to
finish, while promise B was waiting for promise A to finish, which caused a deadlock.

The evidence for this was that when DANBOORU_MAX_CONCURRENCY was lowered from 16 to 8, and therefore
the thread pool size was lowered to 8, it greatly increased the frequency of stuck uploads because
it was more likely for promises to get queued up in a way that caused deadlock.

The problem can be demonstrated with this code:

    # This would deadlock before this commit, but not after.
    128.times.parallel_each do
      8.times.parallel_each do
        4.times.parallel_each do
          # do nothing
        end
      end
    end

The fix is to disable queuing of tasks in thread pools. Instead, if you try to create a promise and
there are no free threads to work it, then it gets worked immediately by the caller itself. This way
it applies backpressure so that a caller can't generate more promises than we have threads,
therefore promises can't be queued up in a way that would cause deadlock.

http://asznajder.github.io/thread-pool-induced-deadlocks/
https://www.springcloud.io/post/2022-04/spring-threadpool/