tags: do large category changes in background job.
Change it so that if an admin or builder manually changes the category of a large tag, it will
update the `posts.tag_count_{category}` fields in a background job instead of in the foreground
request. Fixes requests timing out when trying to change the category of a sufficiently large tag.

Also change it so that tag counts are updated in parallel so that category changes are faster.

Lastly, fix it so that the post is locked while its category counts are being updated, to prevent
the post's tags from being changed while its counts are being updated.