docker: move gems and node_modules out of /danbooru folder.
Fix it so that Ruby gems are installed at /home/danbooru/bundle instead of /danbooru/vendor, and
Javascript packages are installed at /node_modules instead of /danbooru/node_modules.

Keeping these things under /danbooru meant that we had to prevent them from being overridden when
code from the host machine was bind-mounted into the container at /danbooru. We mounted anonymous
volumes to avoid this, but this caused other problems: Docker had to copy files from the container
to the volume each time a container was booted, which was slow, and also it didn't work in Github
Codespaces since there the files aren't copied for some reason.

Moving /danbooru/node_modules to /node_modules works because Node searches for node_modules up the
directory tree until it finds it. We do it this way because there's no way to put it in
/home/danbooru and just tell Node to the path to the `node_modules` folder.

This also fixes it so that you can upgrade gems in the devcontainer with `bundle install` without
having to do `unset BUNDLE_DEPLOYMENT` first.