media assets: add SFTP storage manager.
Add back the SFTP storage manager for storing media assets. Reverts 6fc0854b.

We used to have this but it was removed in favor of sshfs. The problem with sshfs is that whenever
there's even a brief network drop, sshfs hangs and anything that even looks at the mountpoint will
hang too. Even doing `ls` or `df` to check the mountpoints will cause these programs to permanently
hang. Even SIGKILL can't kill them once they get stuck like this. Only a reboot can clear them. This
leads to all webserver and job worker processes eventually getting hung and the site going down
until the servers are rebooted.

`rclone mount` has the same problem, so it's probably a problem with FUSE itself rather than sshfs.

The reason we switched to sshfs in the first place was because our SFTP implementation was too slow.
It was slow because it established a new SFTP connection on every operation. We avoid that now by
using a connection pool to reuse open connections.

We take extra care to try to handle network failures gracefully. If there's a network drop, then
operations will block until the network is back or failed keepalives cause it to time out. If an
operation raises an unexpected exception, then the whole connection pool will be purged and reloaded
just in case one of the connections has gone bad.