docker: fix bin/dev taking a long time to stop.
Fix `bin/dev up` taking 10 seconds to stop after you hit Ctrl-C.

This was because SIGTERM wasn't getting properly forwarded in the webpack container, so webpack
didn't stop until Docker timed out and forcibly killed it.

When you hit Ctrl-C, Docker sends SIGTERM to PID 1 in the container. The process tree of the webpack
container looks like this:

    tini -- bin/shakapacker-dev-server
      npm exec webpack serve --config /danbooru/config/webpack/webpack.config.js --progress --color --no-hot
        sh -c webpack serve --config /danbooru/config/webpack/webpack.config.js --progress --color --no-hot
          webpack

So `tini` receives SIGTERM, which forwards it to `npm`, which forwards it to `sh`, but `sh` doesn't
forward it to `webpack`, so `webpack` never stops and Docker has to kill it after a timeout.

Using `tini -g` makes `tini` forward signals to all subprocesses in its child process group instead of
just its direct children, so `webpack` gets the signal and stops.

https://github.com/krallin/tini#process-group-killing