autocomplete: fix bug with selecting menu items.
Fix a bug with autocomplete where if you selected an entry from the autocomplete menu using the mouse,
then it wouldn't trigger an input event on the input field. This could cause the forgot password
form to be unsubmittable because it wasn't recognized that the username field was filled out.

Also fix a bug where on certain input fields, if you pressed tab to select an item from the autocomplete
menu, then the menu wouldn't be closed afterwards. Selecting the item with the mouse worked correctly.
This happened on fields that accepted a single tag name, artist name, user name, etc.