rate limits: limit IPs by /24 or /64 subnets.
Change it so that IP-based rate limits are based on the /24 (for IPv4) or /64 (for IPv6) subnet
instead of on the individual IP.

This fixes it so that if you have multiple IPs on the same subnet, you can't just rotate between
them to get around rate limits. This was especially easy with IPv6, since normally users are
assigned a /64 which gives them a practically unlimited supply of unique IPs.