codespaces: fix `HTTP Origin header didn't match request.base_url` error.
Fix `HTTP Origin header didn't match request.base_url` errors when running in Github Codespaces and
trying to create an account, login, or do anything else that required a POST request.

Normally when you send a POST request, the browser sends an Origin header containing the URL of the
site sending the request. Rails compares the Origin header with the X-Forwarded-Host header to make
sure the request came from the same site, to prevent CSRF attacks. In Codespaces, the site URL will
be something like https://expert-spoon-abc123-3000.app.github.dev. Codespaces will send that in the
X-Forwarded-Host header, but not in the Origin header. Instead of passing along the Origin header
from the browser, it forcibly overrides it to be https://localhost:3000. This breaks the Origin
check performed by Rails.

The fix is to disable the HTTP Origin check in Codespaces. This is safe because we have other ways
to prevent CSRF attacks (namely CSRF tokens, and the SameSite=Lax flag on cookies).