codespaces: fix passing X-Forwarded-Host header.
Fix it so that we pass along the X-Forwarded-Host header when running behind a reverse proxy.

This fixes Rails returning broken redirects when running in Github Codespaces.

When an action returns a redirect to a relative URL (for example, `redirect_to posts_path`), Rails
converts the relative URL to an absolute URL for the Location header. To generate absolute URLs,
Rails uses the X-Forwarded-Host header to find out the hostname, or the Host header if X-Forwarded-For
isn't present. When running in Codespaces, Github sends `localhost` in the Host header even when the
actual hostname in the X-Forwarded-Host header is something like `expert-spoon-abc123-3000.app.github.dev`.
This meant we generated redirects to localhost instead of to the Github URL when X-Forwarded-Host
wasn't present.

This also adds a /nginx/headers endpoint so we can see the headers received by Nginx for debugging.