devcontainers: allow using docker inside devcontainers.
Give the devcontainer access to the Docker daemon on the host. This lets you run Docker commands
inside the devcontainer.

This effectively gives the devcontainer root access to the host machine, since giving access to
Docker is effectively giving root. This is considered safe because the only things that have access
to the devcontainer are VS Code and the user themselves, and the user already has Docker access on
the host machine. Danbooru itself runs in a separate container from the devcontainer, so an exploit
in Danbooru itself wouldn't give access to the devcontainer. The main risk here is that an exploit
in a VS Code extension could lead to root access to the host machine - but that's also true when
running VS Code outside of a devcontainer.

This is necessary because the only way to do things like view Docker logs or run commands in other
containers from inside a devcontainer is to give it Docker access.