config: change default database name from `danbooru2` to `danbooru`.
Change the default database name from `danbooru2` to `danbooru`.

If you're running Danbooru in Docker Compose, you don't have to do anything. The default database
name in Docker Compose was already `danbooru`.

If you're running a manual install of Danbooru, you will have to either rename your database, or
change your Danbooru config to use the old name.

To rename your database, do:

   bin/rails dbconsole -p
   ALTER DATABASE danbooru2 RENAME TO danbooru;

To change your config to use the old name, set DATABASE_URL in .env.local, like this:

  DATABASE_URL=postgresql://danbooru@localhost/danbooru2

If that doesn't work, you can edit config/database.yml and change it back to what it was before.

This is being changed so that we use the same database name both inside and outside of Docker.

This also fixes an issue with running the tests inside Docker. Before if you ran `bin/rails test`
inside Docker Compose, then it would use the same database for both the tests and for development.
This would trigger a warning from Rails and would break tests if your development database was
populated. This happened because the database name was hardcoded inside DATABASE_URL. The fix is to
not hardcode the database name inside the DATABASE_URL, but that meant we defaulted to `danbooru2`
instead of `danbooru`, hence why we need to change the default.

(The reason the database was called `danbooru2` to begin with was because Danbooru was rewritten
around 2013 and that's what the new database was called to avoid conflicting with the old database).