bin/dev: detect reverse proxies automatically.
Fix it so that you don't have to set `DANBOORU_REVERSE_PROXY=true` manually when running `bin/dev`
behind a reverse proxy such as Ngrok.

DANBOORU_REVERSE_PROXY controls whether the X-Forwarded-* headers sent by the client should be
trusted. When DANBOORU_REVERSE_PROXY is true, then we assume we may be behind a reverse proxy and any
X-Forwarded-* headers (if present) are passed upstream to Danbooru. When it's false, we assume we're
not behind a reverse proxy and any X-Forwarded-* headers set by the client are ignored.

If DANBOORU_REVERSE_PROXY is true and we're not behind a reverse proxy, then it's possible for users
to spoof the X-Forwarded-* headers to spoof their IP or the server hostname. When running under
bin/dev, DANBOORU_REVERSE_PROXY defaults to true because we assume dev instances aren't exposed to
untrusted users. When running in production mode with `docker compose up`, DANBOORU_REVERSE_PROXY
defaults to false because we assume production instances may be exposed to untrusted users.