posts: rework ugoira player.
Rewrite the ugoira player to make it look and feel more like native video players:

* Make the playback bar appear as an overlay when you hover over the ugoira.
* Show the current time and video duration in the playback bar.
* Allow clicking the ugoira to play or pause it.
* Allow viewing the ugoira in fullscreen mode.
* Add keyboard shortcuts: space to pause, left and right to rewind or fast forward.
* Automatically pause the ugoira when tabbed out to avoid wasting CPU / battery life.
* Fix playback to feel smoother.
* Get rid of the jQuery UI dependencies.

This required rewriting it so that the actual animation of the ugoira is handled by us. The only
thing the pixiv-ugoira-player.js code is used for now is loading the ugoira, not for playback.

The main problem with the previous player was that it was frame-based, not time-based, so it only
kept track of the current frame, not the current time. This made seeking to a specific point in time
rather than a specific frame difficult, and it made playback feel jerky. Ugoiras can have a variable
frame rate, and some artists (ab)use this by making animations with frames that pause for several
seconds. This made the playback bar feel jerky, because it could randomly stop and start as you
encountered long frames that paused for several seconds, instead of progressing smoothly as time passed.

The fix was to switch to time-based playback, which means keeping track of the current time and
rendering frames as we go. This makes playback smoother, and it allows you to scrub through the
playback bar smoothly at a constant rate instead of jumping frame-to-frame.