ugoiras: use same video player for original ugoiras and webm samples.
Rewrite the ugoira player to use the same player for both the webm sample and the original ugoira
file. This allows you to seamlessly switch between the sample and the ugoira while the video is
playing, in much the same way you can switch video quality in other video players.

This entailed refactoring things to separate the video player controls from the ugoira renderer so
that we can use the same player for both regular HTML <video>s and ugoiras. The UgoiraRenderer class
emulates the HTMLMediaElement interface so that the player can treat it the same as a normal video.