docker: upgrade ffmpeg to 7.1.1.
This fixes a bug where the duration of APNG images was calculated incorrectly (the last frame wasn't
counted).

This also makes it so that static AVIF images that contain an auxiliary video (aka "live photos")
are now treated as animated files, with the duration and frame count taken from the video, even if
the AVIF file is set to show the image by default. This means these images can't be uploaded because
animated AVIFs aren't allowed.