Fix #6004: Notes can cover media controls on mobile.
Fix it so that it's not possible to add notes to ugoira or video posts. Adding notes to videos and
ugoiras was disallowed before, but due to a regression it became possible to add notes to ugoiras.

Also fix the "Add note" link appearing and doing nothing on posts where it isn't possible to add
notes, and fix it being possible to add notes to video posts through the API.